+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Project: Night festa
 - Project: Hitogomi
 - project: Vanish cat
+- project: Girl hearing a moonlight
 
 ## [0.6.0] - 2019-08-01
 ### Added