+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Project: Forbidden love
 - Project: Blue ghost
 - Project: Night festa
+- project: Girl hearing a moonlight
+
+## [0.7.0] - 2019-09-14
+### Added
 - Project: Hitogomi
 - project: Vanish cat
-- project: Girl hearing a moonlight
 
 ## [0.6.0] - 2019-08-01
 ### Added
@@ -98,7 +101,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Production: slime love story
 - Production: kujo saeko
 
-[Unreleased]: https://github.com/nagisc007/prj_estar/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/nagisc007/prj_estar/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/nagisc007/prj_estar/releases/v0.7.0
 [0.6.0]: https://github.com/nagisc007/prj_estar/releases/v0.6.0
 [0.5.0]: https://github.com/nagisc007/prj_estar/releases/v0.5.0
 [0.4.0]: https://github.com/nagisc007/prj_estar/releases/v0.4.0